.history/scripts/requirements_20240729055852.txt
requests
pandas
matplotlib
seaborn
yfinance
python-dotenv
